moved createServiceMember to support folder for re-use
